feat: Application.kt install content negotiation
To enable content negotiation, we need to install
the ContentNegotiation feature. This feature is
used to automatically serialize and deserialize
the request and response bodies.
The ContentNegotiation feature is installed by
calling the install(ContentNegotiation) method on
the application object.

We include json and protobuf as the supported
formats for serialization and deserialization.

Signed-off-by: gilangrilhami <gilangrilhami@hotmail.co.uk>