feat: POST endpoint for OpenTelemetry trace data
This commit adds a POST endpoint to the application
that accepts OpenTelemetry trace data. The endpoint
is defined in the `Routing.kt` file and is implemented
in the `Application.kt` file. The endpoint accepts
a JProtobuf-encoded data structure that represents
a list of OpenTelemetry spans. The data is decoded
and processed by the application.

Endpoint is `POST /v1/traces`.

Signed-off-by: gilangrilhami <gilangrilhami@hotmail.co.uk>